requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
dataclasses==0.8
idna==2.10
idna-ssl==1.1.0
multidict==5.0.2
PythonTwitchBotFramework
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
websockets==8.1
yarl==1.6.3
